Rename find_config to cyhy_config
